fix(worktree): move PathTarget to leaf pathUtils and fix prop name mismatch
decision(filer-recursive): FileTreeItem の再帰呼び出し点 :selected-path="selectedRelPath" を :selected-rel-path に修正。前 commit で identifier は replace_all したが kebab-case の attribute 名を取りこぼし、深さ 1 以降の全ノードで子 prop が undefined になり選択ハイライト (bg-zinc-700 / text-white) が機能していなかった退行。vue-tsc / eslint / bun test では捕捉されない領域
decision(path-target-leaf): PathTarget 型と pathTargetToString 関数を useWorktreeStore.ts から pathUtils.ts (副作用なし leaf) に移動。useWorktreeStore.ts に置くと bun:test で worktree barrel ロード経由で window.__gozdReceive 副作用が走る制約があったが、leaf 移動 + bun-test-setup の window stub で構造的に解消し、findAbsolutePathMatches.ts から targetDisplay 複写を削除して関数 SSOT を完全に確立
decision(bun-test-setup): bunfig.toml + bun-test-setup.ts を追加し globalThis.window を最小スタブする。shared/rpc/messages.ts のモジュールトップレベル副作用 `window.__gozdReceive = ...` が DOM 不在の bun:test で ReferenceError になっていた根本原因を消し、worktree barrel を test 環境から import できる構造に
decision(normalize-relative-doc): pathUtils.ts:normalizeRelative の jsdoc に「~ を通常 segment として保持する (escapesWorktree が後段で home 参照として弾く前提)」契約を明記。将来 normalize 実装が tilde 展開する変更を入れた瞬間に escapesWorktree の前提が壊れる不変条件を文書化
constraint(bun-test): bun:test は DOM globals を持たないため shared/rpc/messages.ts の window 参照が必ず走る。preload 経由のスタブが唯一の構造的対応 (個別 mock より test 全体の前提として明示的)
