+++ apps/renderer/bunfig.toml
@@ -1,0 +1,2 @@
+[test]
+preload = ["./bun-test-setup.ts"]
