+++ apps/renderer/src/features/worktree/index.ts
@@ -1,10 +1,11 @@
 export { UNCOMMITTED_HASH } from "./constants";
 export { generateTimestamp } from "./generateTimestamp";
-export { normalizeAbsolute, normalizeRelative } from "./pathUtils";
+export { normalizeAbsolute, normalizeRelative, pathTargetToString } from "./pathUtils";
+export type { PathTarget } from "./pathUtils";
 export type { GitStatusChangePayload } from "./rpc";
 export { default as StatusIcons } from "./StatusIcons.vue";
-export { pathTargetToString, useWorktreeStore } from "./useWorktreeStore";
-export type { PathTarget, Selection } from "./useWorktreeStore";
+export { useWorktreeStore } from "./useWorktreeStore";
+export type { Selection } from "./useWorktreeStore";
 export { useGitStatusStore } from "./useGitStatusStore";
 export { useGitStatusSync } from "./useGitStatusSync";
 export { useRemoteFetchSync } from "./useRemoteFetchSync";
